+++ lib/ghamma/cli.rb
@@ -23,9 +23,11 @@ module Ghamma
       argument :owner, required: true, desc: "The user or organisation to whom the repo belongs"
       argument :repo, required: true, desc: "The repo to whom the workflow belongs"
       argument :workflow, required: true, desc: "The filename for the desired workflow, e.g. tests.yml"
+      option :since, default: "1970-01-01", desc: "Fetch workflow runs since this date"
 
-      def call(owner:, repo:, workflow:)
-        durations = GithubApiClient.new(owner: owner, repo: repo, token: ENV["GH_TOKEN"]).fetch_workflow_duration_history(workflow)
+      def call(owner:, repo:, workflow:, since:)
+        durations = GithubApiClient.new(owner: owner, repo: repo, token: ENV["GH_TOKEN"])
+          .fetch_workflow_duration_history(workflow, since)
 
         durations_output = CSV.generate do |csv|
           csv << ["Date", "Duration"]